go/netwatch: add initial version of netwatch tool
At a high level, we wish to track the presence of devices on a LAN. We
use the bettercap for most of the heavy lifting. That tool doesn't keep
track of state, so we do so here in this command (albeit only in memory
now (albeit only in memory now).

More specifically we initially only do the following:
- Create a new bettercap session.
- Start the net.probe module.
- Scan for new endpoints found on session's LAN.
- Log out new, dropped, and returned endpoints.