go/netwatch: remove Diff struct, and create TTL expire callbacks
There was a tight coupling between the Watcher.Respond logging and the
diff construction. It was getting awkward so we just smooshed them
together.

In the process we also fix some reported timing bugs, and make the
check.sh script fail on first error.

The complexity of the event loop is getting so complicated as to really
necessitate tests for productive future development. Let's start
exploring the reading packets from offline files!